backend tests - auth & quote routes
